:man_shrugging: Remove parseValues as I implemented it wrong
I had used parseValue as a flag to enable parsing of nested keys which is
not what this option value was ment for. This option value comes from nconf
and is used to parse known values from string to their proper types. The
parseing of nexted values just happens by default.