Rename logicalSeparator to simply separator
The env() store which I use mostly calls the options property to define
the nested separator well 'separator' however when I was implementing this
pluging I extended the ncof.Memory store which uses the property
'logicalSeparator' instead. As I will likly use this plugin with both
an env _and_ secret store and have both use the same seperator it would
be nice if they both used the same options property to define the
separator character.

Since however I'm not overriding the get method from ncof.Memory I still
need to define the logicalSeparator property.

This change set updates the plugin to take in a 'separator' property,
so the public API will be 'separator', but uses 'logicalSeparator'
internally.