Add parseValues option
This change set implements the ability for this plugin to properly parse
well known values obtained from the secrets store. Just like with the
env store if the secret is a string 'false' when this option is set the
get method will return the value false vs. the string false automatically.